backend/requirements.txt
Flask
psycopg2-binary
faker
flasgger
flask_cors
